Remove explanation from non-ConceptMap answers types
