+++ concept/answer/Answer.java
@@ -19,33 +19,9 @@
 
 package grakn.core.concept.answer;
 
-import javax.annotation.CheckReturnValue;
-import javax.annotation.Nullable;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.Set;
-
 /**
  * An object that contains the answer of every Graql Query.
  */
 public abstract class Answer {
 
-    /**
-     * @return an explanation object indicating how this answer was obtained
-     */
-    @Nullable
-    @CheckReturnValue
-    public abstract Explanation explanation();
-
-    /**
-     * @return all explanations taking part in the derivation of this answer
-     */
-    @CheckReturnValue
-    public Set<Explanation> explanations() {
-        if (this.explanation() == null) return Collections.emptySet();
-        Set<Explanation> explanations = new HashSet<>();
-        explanations.add(this.explanation());
-        this.explanation().getAnswers().forEach(ans -> explanations.addAll(ans.explanations()));
-        return explanations;
-    }
 }
