+++ concept/answer/ConceptList.java
@@ -31,21 +31,11 @@ public class ConceptList extends Answer {
 
     // TODO: change to store List<Concept> once we are able to construct Concept without a database look up
     private final List<ConceptId> list;
-    private final Explanation explanation;
 
     public ConceptList(List<ConceptId> list) {
-        this(list, new Explanation());
-    }
-
-    public ConceptList(List<ConceptId> list, Explanation explanation) {
         this.list = Collections.unmodifiableList(list);
-        this.explanation = explanation;
     }
 
-    @Override
-    public Explanation explanation() {
-        return explanation;
-    }
 
     public List<ConceptId> list() {
         return list;
